Replace SPI-based implementation with shaded jar for Calcite
