--- cdap-relational-spi/src/main/java/io/cdap/cdap/etl/spi/relational/SQLDialectException.java
@@ -1,30 +1,0 @@
-/*
- * Copyright © 2023 Cask Data, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-
-package io.cdap.cdap.etl.spi.relational;
-
-/**
- * This exception is thrown when a problem is encountered by the SQL dialect conversion process.
- */
-public class SQLDialectException extends Exception {
-    public SQLDialectException(String message) {
-        super(message);
-    }
-
-    public SQLDialectException(String message, Throwable exception) {
-        super(message, exception);
-    }
-}
